Add persistent page view tracking
Use the browser's built-in localStorage object to track the pages
that a user of the site visits.

Pages are tracked even if they have already been added to the
list of pages. The idea is to have a list of urls in the order
that the user traversed through the website. For example:

page_views = [url1, url2, url3, url2, url3, ...]